Initial commit, first release, RAF Technical Note on Processing Algorithms
Contains descriptions of the algorithms used to calculate variable that
are included in NCAR/EOL/RAF data files. This is the first nearly
complete version, but there is still a list of desirable changes that
is appended to the document (ProcessingAlgorithms.pdf). This technical
note is expected to change frequently as changes are made to the processing
algorithms, so there is also a description of how to produce the document
from the files in this repository.